requirements.txt
# Requirements to build and test
google-cloud-datastore==1.7.1
